Generify internal type of Generator
Allows user to specify the internal type yielded by the Generator. This 
is within the bounds of the FromIterator<u8> trait, meaning as long as 
the collection can be formed from an Iterator over <u8>, the it can be 
used with Generator. This type is specified using "turbofish" or normal 
type annotations.